requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
filelock==3.11.0
Pillow==9.5.0
platformdirs==3.2.0
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0
